requirements.txt
absl-py
astor
astunparse
cached-property
cachetools
certifi
charset-normalizer
clang
cycler
dataclasses
flatbuffers
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
imageio
importlib-metadata
ipython
jedi
Keras
Keras-Applications
Keras-Preprocessing
kiwisolver
libclang
Markdown
matplotlib
matplotlib-inline
mock
networkx
numpy
oauthlib
opencv-python
opt-einsum
packaging
parso
Pillow
prompt-toolkit
protobuf
pyasn1
pyasn1-modules
pydicom
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
pylibjpeg-rle
pyparsing
python-dateutil
PyWavelets
PyYAML
requests
requests-oauthlib
rsa
scikit-image
scipy
six
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
termcolor
tifffile
typing-extensions
urllib3
Werkzeug
wincertstore
wrapt
zipp
